Added win and added greater speed and obstacles.

+++ SDLTest-master/SDLGameTest/game/TextureMgr.h
@@ -16,7 +16,7 @@
 class TextureMgr  
 {
 public:
-	enum TextureID: size_t {
+	enum TextureID: uint32_t {
 		ENEMY,
 		PLAYER,
 		PLAYER_BULLET,
